requirements.txt
fastapi
uvicorn
feedparser
newspaper3k
sumy
lxml
lxml_html_clean
nltk